refactor: cache key
