+++ packages/backend/src/app/workers/sandbox/provisioner/sandbox-cache-key.ts
@@ -1,0 +1,77 @@
+import { FileId, FlowVersionId, apId } from '@activepieces/shared'
+import { logger } from '../../../helper/logger'
+
+export enum SandBoxCacheType {
+    CODE = 'CODE',
+    FLOW = 'FLOW',
+    NONE = 'NONE',
+    PIECE = 'PIECE',
+}
+
+
+export type TypedProvisionCacheInfo<T extends SandBoxCacheType = SandBoxCacheType> =
+    T extends SandBoxCacheType.CODE
+        ? CodeProvisionCacheInfo
+        : T extends SandBoxCacheType.FLOW
+            ? FlowProvisionCacheInfo
+            : T extends SandBoxCacheType.NONE
+                ? NoneProvisionCacheInfo
+                : T extends SandBoxCacheType.PIECE
+                    ? PieceProvisionCacheInfo
+                    : never
+
+export type ProvisionCacheInfo = TypedProvisionCacheInfo
+
+export const extractProvisionCacheKey = (params: ProvisionCacheInfo): string => {
+    logger.debug({ type: params.type }, '[SandboxProvisioner#extractCacheKey]')
+    switch (params.type) {
+        case SandBoxCacheType.CODE:
+            return extractCodeCacheKey(params)
+        case SandBoxCacheType.FLOW:
+            return extractFlowCacheKey(params)
+        case SandBoxCacheType.NONE:
+            return extractNoneCacheKey(params)
+        case SandBoxCacheType.PIECE:
+            return extractPieceCacheKey(params)
+    }
+}
+
+
+
+const extractCodeCacheKey = ({ artifactSourceId }: CodeProvisionCacheInfo): string => {
+    return `CODE-artifactSourceId-${artifactSourceId}`
+}
+
+const extractFlowCacheKey = ({ flowVersionId }: FlowProvisionCacheInfo): string => {
+    return `FLOW-flowVersionId-${flowVersionId}`
+}
+
+const extractNoneCacheKey = (_params: NoneProvisionCacheInfo): string => {
+    return `NONE-apId-${apId()}`
+}
+
+const extractPieceCacheKey = ({ pieceName, pieceVersion }: PieceProvisionCacheInfo): string => {
+    return `PIECE-pieceName-${pieceName}-pieceVersion-${pieceVersion}`
+}
+
+
+type BaseProvisionCacheInfo<T extends SandBoxCacheType> = {
+    type: T
+}
+
+type CodeProvisionCacheInfo = BaseProvisionCacheInfo<SandBoxCacheType.CODE> & {
+    artifactSourceId: FileId
+}
+
+type FlowProvisionCacheInfo = BaseProvisionCacheInfo<SandBoxCacheType.FLOW> & {
+    flowVersionId: FlowVersionId
+}
+
+type NoneProvisionCacheInfo = BaseProvisionCacheInfo<SandBoxCacheType.NONE>
+
+type PieceProvisionCacheInfo = BaseProvisionCacheInfo<SandBoxCacheType.PIECE> & {
+    pieceName: string
+    pieceVersion: string
+}
+
+
