--- packages/backend/src/app/workers/sandbox/provisioner/sandbox-cache-type.ts
@@ -1,6 +1,0 @@
-export enum SandBoxCacheType {
-    CODE = 'CODE',
-    FLOW = 'FLOW',
-    NONE = 'NONE',
-    PIECE = 'PIECE',
-}
