+++ packages/backend/src/app/workers/sandbox/provisioner/sandbox-provisioner.ts
@@ -1,20 +1,14 @@
 import { Sandbox } from '..'
 import { sandboxCachePool } from '../cache/sandbox-cache-pool'
 import { sandboxManager } from '../sandbox-manager'
-import { ApEnvironment, FileId, FlowVersionId, apId } from '@activepieces/shared'
-import { SandBoxCacheType } from './sandbox-cache-type'
+import { FileId } from '@activepieces/shared'
+import { SandBoxCacheType, TypedProvisionCacheInfo } from './sandbox-cache-key'
 import { logger } from '../../../helper/logger'
-import { system } from '../../../helper/system/system'
-import { SystemProp } from '../../../helper/system/system-prop'
 
 export const sandboxProvisioner = {
     async provision({ pieces = [], codeArchives = [], ...cacheInfo }: ProvisionParams): Promise<Sandbox> {
-        const cacheKey = extractCacheKey(cacheInfo)
 
-        const cachedSandbox = await sandboxCachePool.findOrCreate({
-            key: cacheKey,
-            type: cacheInfo.type,
-        })
+        const cachedSandbox = await sandboxCachePool.findOrCreate(cacheInfo)
 
         await cachedSandbox.prepare({
             pieces,
@@ -24,7 +18,7 @@ export const sandboxProvisioner = {
         const sandbox = await sandboxManager.allocate()
 
         await sandbox.assignCache({
-            cacheKey,
+            cacheKey: cachedSandbox.key,
             cachePath: cachedSandbox.path(),
         })
 
@@ -44,84 +38,20 @@ export const sandboxProvisioner = {
     },
 }
 
-const extractCacheKey = (params: ProvisionCacheInfo): string => {
-    logger.debug({ type: params.type }, '[SandboxProvisioner#extractCacheKey]')
-    const environment = system.get(SystemProp.ENVIRONMENT)
-    if (environment === ApEnvironment.DEVELOPMENT) {
-        // In development, we don't want to cache anything to prevent issue with linking packages in hot reload
-        return extractNoneCacheKey({ type: SandBoxCacheType.NONE })
-    }
-    switch (params.type) {
-        case SandBoxCacheType.CODE:
-            return extractCodeCacheKey(params)
-        case SandBoxCacheType.FLOW:
-            return extractFlowCacheKey(params)
-        case SandBoxCacheType.NONE:
-            return extractNoneCacheKey(params)
-        case SandBoxCacheType.PIECE:
-            return extractPieceCacheKey(params)
-    }
+type ProvisionParams<T extends SandBoxCacheType = SandBoxCacheType> = TypedProvisionCacheInfo<T> & {
+    pieces?: Piece[]
+    codeArchives?: CodeArchive[]
 }
 
-const extractCodeCacheKey = ({ artifactSourceId }: CodeProvisionCacheInfo): string => {
-    return `CODE-artifactSourceId-${artifactSourceId}`
-}
-
-const extractFlowCacheKey = ({ flowVersionId }: FlowProvisionCacheInfo): string => {
-    return `FLOW-flowVersionId-${flowVersionId}`
-}
-
-const extractNoneCacheKey = (_params: NoneProvisionCacheInfo): string => {
-    return `NONE-apId-${apId()}`
-}
-
-const extractPieceCacheKey = ({ pieceName, pieceVersion }: PieceProvisionCacheInfo): string => {
-    return `PIECE-pieceName-${pieceName}-pieceVersion-${pieceVersion}`
-}
 
 type Piece = {
     name: string
     version: string
 }
 
-type BaseProvisionCacheInfo<T extends SandBoxCacheType> = {
-    type: T
-}
-
-type CodeProvisionCacheInfo = BaseProvisionCacheInfo<SandBoxCacheType.CODE> & {
-    artifactSourceId: FileId
-}
-
-type FlowProvisionCacheInfo = BaseProvisionCacheInfo<SandBoxCacheType.FLOW> & {
-    flowVersionId: FlowVersionId
-}
-
-type NoneProvisionCacheInfo = BaseProvisionCacheInfo<SandBoxCacheType.NONE>
-
-type PieceProvisionCacheInfo = BaseProvisionCacheInfo<SandBoxCacheType.PIECE> & {
-    pieceName: string
-    pieceVersion: string
-}
-
-type ProvisionCacheInfo<T extends SandBoxCacheType = SandBoxCacheType> =
-    T extends SandBoxCacheType.CODE
-    ? CodeProvisionCacheInfo
-    : T extends SandBoxCacheType.FLOW
-    ? FlowProvisionCacheInfo
-    : T extends SandBoxCacheType.NONE
-    ? NoneProvisionCacheInfo
-    : T extends SandBoxCacheType.PIECE
-    ? PieceProvisionCacheInfo
-    : never
-
 type CodeArchive = {
     id: FileId
     content: Buffer
-}
-
-type ProvisionParams<T extends SandBoxCacheType = SandBoxCacheType> = ProvisionCacheInfo<T> & {
-    pieces?: Piece[]
-    codeArchives?: CodeArchive[]
 }
 
 type ReleaseParams = {
